fix:server.js and server2.js
